requirements.txt
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-apscheduler==0.6.2
django-environ==0.8.1
django-qr-code==3.0.0
djangorestframework==3.13.1
emoji==1.6.1
gunicorn==20.1.0
kombu==5.2.4
loguru==0.5.3
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==3.0.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
rsa==4.8
segno==1.4.1
six==1.16.0
sqlparse==0.4.2
Telethon==1.24.0
tzdata==2022.1
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
